add a link API to graphCache
This allows the user to pass in a rootKey or entity from which we'll derive the rootKey for instance Query or { __typename:'Todo', id: 1 }, secondly we accept a field, this can be for instance author which will be applied to the rootKey. As a thirth argument we accept the entity or entities to be written as keys to the rootKey.field, for example { __typename: 'Author', id: 1 } and as a last argument we accept fieldArguments, these will be applied to the fieldKey and will help with fields that have their own arguments for instance: Todo:1.authors({ first: 10 }).